components_service/requirements.txt
grpcio-tools~=1.30
bcrypt
sqlalchemy >= 2.0.25
psycopg2-binary